Port to RISC-V (RV64GC)
** This change wrote written and debugged entirely on a (virtual)
   RISC-V machine running Fedora/RISC-V **

Surprisingly, this port was quite straight forward, so I went ahead
and made the codegen pretty decent (and in the process made it
slightly more complicated).

This clearly illustrates how RISC[-V] makes for a nicer compilation
target.

Example: for (1 + x*3 + 4*(5 + y)) * (1 + x*3 + 4*(5 + y))

     0x200001f000:        lw      t0,480(a0)
     0x200001f004:        li      t1,3
     0x200001f008:        mul     t2,t0,t1
     0x200001f00c:        lw      s0,484(a0)
     0x200001f010:        li      s1,4
     0x200001f014:        mul     a1,s0,s1
     0x200001f018:        li      a2,21
     0x200001f01c:        add     a3,a1,a2
     0x200001f020:        add     a4,t2,a3
     0x200001f024:        mul     a0,a4,a4
     0x200001f028:        ret

NB: the usage tracking is broken so when there are common shared
expressions, we don't free all the registers we should.  Fix it later.